feat(wallet): configure rewards per source on the Configuration page
Puts the reward settings back where they were expected — on Approbations
> Wallet > Configuration — now with a section per source: épreuve,
examen national, concours. The examen national block is the one that
never existed; épreuve and concours had cards there before, bound to the
global rewardPerExam / rewardPerConcours this branch removes.

Each card reads and writes its own row through
/user-payment/admin/reward-configurations/:sourceType, so it carries the
full per-source settings — amount, review delay, admin-validation and
per-source kill switch, plus the four quotas — and saves on its own
button, matching the one-PATCH-per-source API.

The standalone /admin/wallet-recompenses page added in the previous
commit is removed along with its route and sidebar entry: keeping both
would mean two places editing the same amount. The UI now lives in
components/RewardConfigurationSection so the config page stays readable.

Quotas keep their explicit "0 = illimité" wording on every field —
read as "no reward", it silently stops paying contributors.

Rendered and checked against the running dev stack: the three cards show
100 / 100 / 200 XOF, above Retraits, with Wallet back to two sidebar
entries.
